+++ CotEditor/Sources/CEAppDelegate.m
@@ -160,7 +160,6 @@
                                                CEDefaultSmartInsertAndDeleteKey: @NO,
                                                CEDefaultEnableSmartQuotesKey: @NO,
                                                CEDefaultEnableSmartDashesKey: @NO,
-                                               CEDefaultEnableAutoSpellingCorrectionKey: @NO,
                                                CEDefaultEnableSmartIndentKey: @YES,
                                                CEDefaultAutoLinkDetectionKey: @NO,
                                                CEDefaultAppendsCommentSpacerKey: @YES,